Set l:job.errorformat before starting the job
Store the current &errorformat with the job (if not already set).
Otherwise it is possible for the wrong errorformat to be applied to job
results.

For example, start a long running compilation using &makeprg. While it
is running, navigate to a buffer with a different errorformat, or run a
quick :grep which leaves the focus on the quickfix with a different
errorformat. Now when the compilation finishes, it tries to apply the
current &errorformat instead of the original.