Prevent expand() from expanding $*
Rather annoyingly, the all-encompassing expand() was expanding $* in
makeprg before it had the chance to be substituted by l:args.

To fix this, s:escape() is now used *after* the `if l:cmd =~# '\$\*'`
block, to make sure $* is substituted by l:args before any expanding
happens.

A more restricted s:specialexpand() is now also used in s:escape(), in
the same vein as s:fnameexpand(). This restricts expansion to
cmdline-special variables such as <cword>, to prevent any runaway
expansion e.g. with globs or shell variables.

Additionally, this allows s:escape() to be applied only once, as opposed
to applying it to all l:args with map() and to l:cmd.